fix(playback): load PCM into xaak after CERTIFIED + E2E verified
Root cause: PCM never loaded into PlaybackEngine after mastering.
play() found no kernel → silent fail → telemetry_worker never spawned.

Fix: DspOutput carries AudioChunk → master.rs calls playback.load()
after CERTIFIED.

E2E proof (--release mode):
  Mastering:  CERTIFIED in ~80s
  Playback:   is_playing=true, duration=158639ms, 48kHz stereo
  UDP:        147514 bytes in 10s ✅
  Worker:     spawned, 50+ chunks at 2048 frames

Full telemetry bridge proven end-to-end:
  xaak audio callback → ringbuf → telemetry_worker
  → bincode → UDP 127.0.0.1:9000 → src-tauri listener
  → Arc<Mutex<RealtimeFrame>> → Tauri command → Dioxus

INV-TB-1..8 certified in production ✅

+++ lineos/m0/m0-daemon/src/handlers/export.rs
@@ -199,7 +199,7 @@ fn export_wav(blob: &StoredBlob, path: &Path) -> Result<(), String> {
     let samples = pcm_bytes_to_f32(&blob.audio_bytes);
 
     let spec = hound::WavSpec {
-        channels:        blob.channels as u16,
+        channels:        blob.channels,
         sample_rate:     blob.sample_rate,
         bits_per_sample: 32,
         sample_format:   hound::SampleFormat::Float,
@@ -242,7 +242,7 @@ fn export_aiff(blob: &StoredBlob, path: &Path) -> Result<(), String> {
     }
 
     let pcm = pcm_bytes_to_f32(&blob.audio_bytes);
-    let channels    = blob.channels.max(1) as u16;
+    let channels    = blob.channels.max(1);
     let sample_rate = blob.sample_rate;
     let num_frames  = (pcm.len() / channels as usize) as u32;
     let bit_depth: u16 = 32;
